code/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.6.0
anyio==4.4.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
billiard==4.2.1
black==23.7.0
camel-converter==4.0.1
celery==5.5.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==41.0.2
Django==4.2.3
django-admin-autocomplete-filter==0.7.1
django-admin-rangefilter==0.12.4
django-admin-sortable2==2.1.4
django-celery-beat==2.7.0
django-cors-headers==3.14.0
django-editor-widgets==4.2
django-ninja==1.1.0
django-timezone-field==7.1
dnspython==2.5.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
flake8==6.0.0
frozenlist==1.4.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
kombu==5.5.2
MarkupSafe==2.1.3
mccabe==0.7.0
meilisearch==0.31.4
multidict==6.0.5
mypy-extensions==1.0.0
ninja==1.11.1.1
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.1
prompt_toolkit==3.0.50
psycopg==3.2.5
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
pyflakes==3.0.1
PyJWT==2.7.0
pymongo==4.6.1
PyMySQL==1.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-engineio==4.5.1
python-socketio==5.8.0
redis==5.0.1
requests==2.31.0
setuptools==70.2.0
six==1.17.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2025.2
Unidecode==1.3.6
urllib3==2.0.3
uvicorn==0.22.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4